quelques bugs

+++ php/controller/ControllerProduit.php
@@ -21,42 +21,71 @@ class ControllerProduit {
 	    }
     }
 
-    public static function create(){    
+    public static function create(){
+        if(Session::is_admin()){
         $v=new ModelProduit("","","");
         $isUpdate=false;
     	$controller='produit'; $view='update'; $pagetitle='creation de produit';     //appel au modèle pour gerer la BD
         require File::build_path('view/view.php');  //"redirige" vers la vue
+        }else{
+            $tab_v = ModelProduit::selectAll();
+            $controller='produit'; $view='list'; $pagetitle='Liste des produits';     //appel au modèle pour gerer la BD
+            require File::build_path('view/view.php');  //"redirige" vers la vue
+        }
     }
 
     public static function created(){
-    	$id=myGet('id');
-    	$nom=myGet('nom');
-    	$prix=myGet('prix');
- 		$v=new ModelProduit($id,$nom,$prix);
- 		$v->save();
-        $controller='produit'; $view='created'; $pagetitle='cree';     //appel au modèle pour gerer la BD
-        $tab_v = ModelProduit::selectAll();
-        require File::build_path('view/view.php');  
+        if(Session::is_admin()){
+        	$id=myGet('id');
+        	$nom=myGet('nom');
+        	$prix=myGet('prix');
+     		$v=new ModelProduit($id,$nom,$prix);
+     		$v->save();
+            $controller='produit'; $view='created'; $pagetitle='cree';     //appel au modèle pour gerer la BD
+            $tab_v = ModelProduit::selectAll();
+        require File::build_path('view/view.php'); 
+        }
+ 
+        else{
+            $tab_v = ModelProduit::selectAll();
+            $controller='produit'; $view='list'; $pagetitle='Liste des produits';     //appel au modèle pour gerer la BD
+            require File::build_path('view/view.php');  //"redirige" vers la vue
+        }
     }
 
     public static function delete(){
+        if(Session::is_admin()){
         $id=myGet('id');
         ModelProduit::delete($id);
         $tab_v=ModelProduit::selectAll();
         $controller='produit'; $view='deleted'; $pagetitle='supprimé';     //appel au modèle pour gerer la BD
         require File::build_path('view/view.php');  
+        }
+        else{
+            $tab_v = ModelProduit::selectAll();
+            $controller='produit'; $view='list'; $pagetitle='Liste des produits';     //appel au modèle pour gerer la BD
+            require File::build_path('view/view.php');  //"redirige" vers la vue
+        }
     }
 
     public static function update(){
+        if(Session::is_admin()){
         $im=myGet('id');
         $v=ModelProduit::select($im);
         $isUpdate=true;
         $controller='produit'; $view='update'; $pagetitle='mise à jour de produit';     //appel au modèle pour gerer la BD
         require File::build_path("view/view.php");  //"redirige" vers la vue
+        }
+        else{
+            $tab_v = ModelProduit::selectAll();
+            $controller='produit'; $view='list'; $pagetitle='Liste des produits';     //appel au modèle pour gerer la BD
+            require File::build_path('view/view.php');  //"redirige" vers la vue
+        }
     }
 
 
     public static function updated(){
+        if(Session::is_admin()){
         $controller='produit'; $view='updated'; $pagetitle='mise à jour de produit';     //appel au modèle pour gerer la BD
         $data=array(
             "id"=>myGet("id"),
@@ -66,6 +95,12 @@ class ControllerProduit {
         ModelProduit::update($data);
         $tab_v = ModelProduit::selectAll();
         require File::build_path("view/view.php");  //"redirige" vers la vue
+        }
+        else{
+             $tab_v = ModelProduit::selectAll();
+            $controller='produit'; $view='list'; $pagetitle='Liste des produits';     //appel au modèle pour gerer la BD
+            require File::build_path('view/view.php');  //"redirige" vers la vue           
+        }
     }
 
     public static function panier(){
